requirements.txt
-e .
aws_cdk.aws_wafv2